+++ packages/core/package.json
@@ -10,7 +10,6 @@
   },
   "dependencies": {
     "@gitbeaker/requester-utils": "^34.2.0",
-    "form-data": "^4.0.0",
     "li": "^1.3.0",
     "mime-types": "^2.1.32",
     "query-string": "^7.0.0",
@@ -20,6 +19,7 @@
     "@rollup/plugin-replace": "^3.0.0",
     "@rollup/plugin-json": "^4.1.0",
     "@types/node": "^16.0.0",
+    "form-data": "^4.0.0",
     "esm": "^3.2.25",
     "fs-extra": "^10.0.0",
     "get-param-names": "github:jdalrymple/get-param-names#1-improve-functionality",
