+++ packages/core/test/unit/infrastructure/Utils.ts
@@ -1,5 +1,4 @@
-import FormData from 'form-data';
-import { getAPIMap, appendFormFromObject } from '../../../src/infrastructure';
+import { getAPIMap } from '../../../src/infrastructure';
 
 jest.mock(
   '../../../dist/map.json',
@@ -14,21 +13,5 @@ describe('getAPIMap', () => {
     expect(() => {
       getAPIMap();
     }).toThrow();
-  });
-});
-
-describe('appendFormFromObject', () => {
-  it('should convert object key/values to formdata instance', () => {
-    const data = { a: 5, b: 'test' };
-    const form = appendFormFromObject(data);
-
-    expect(form).toBeInstanceOf(FormData);
-  });
-
-  it('should convert object key/values with metadata to formdata instance', () => {
-    const data = { a: 5, b: ['test', { filename: 'name.jpg' }] };
-    const form = appendFormFromObject(data);
-
-    expect(form).toBeInstanceOf(FormData);
   });
 });
